Weight SVMs predictions (using scores)
Merge the prediction of all the SVMs (there is one SVM per feature) by weighting them regarding their performance. To do so, the score of each SMV is multiplied by its beta coeficient (its performance) and summed with the result of this calculus per all SVMs. The result is a number between -inf and inf, which is converted in the class (0 if the number is negative or 0, or class 1 elsewise) per each instance.