SVM beta coeficients fixed and MNRML dimensionality correction
The beta coeficients are now used properly (multiplyed per feature and fold, in contrast to using their mean value per feature as it was before).
Moreover, a correction in the dimensionality of the PCA is done in the MNRML projection. Before, the new dimension was automatically calculated by using PCA's eigen vector values' dimension (as default) but it generated errors at rounding.
The string function has been replaced by num2str in order to offer retrocompatibility.