Fixed MNRML classification problems
In all the previous commits, classification with original VGG-Face and VGG-F features worked well, but it did not work at all using MNRML and NRML projections (it was like a coin flip for binary classification).
We have seen that at the time of merging the features of the pair (parent with the child) using the following formula: |v1-v2|/(v1+v2) if the dividend is very small it does not work well because all the values increase a lot.
With these fixes, the classification gets results comparable to the originam MNRML paper.