(not working) no PCA experiment
When calling mnrml_train function, if previously PCA has been applyed to all the features (so that all have the same dimension) it works. However, in this commit we have tryed to remove the PCA and try mnrml with the original features. This raises the following error:

Inner matrix dimensions must agree.

Error in mnrml_train (line 27)
        Xaa = Xa{p} * W;